Global Modules
Separated the modules out into global modules. Each module now declares its
dependencies at the top of the page and in its paramaters. A module wrapper was
written for the utility functions and a mediator was written to manage the
calculator, the point buffer, and the plots. required some rewrites in the
other modules.